removed storyboard added some things
so now the storyboards are disconnected but has tab bar connecting them all programmatically, hopefully this will make it so the storyboards open in Xcode much faster.
not sure what else, oh yea so i made it so the labels in the confirmation page stay in the same position, but the image positions will change. SUSPENSE