updated New dashbaord tiles, and reduced code duplication
the global styles class is now a singleton because of how many instances 
of it i was having to create it just makes sense to have a singleton??

moved the struct of the plist into other views with @State and 
@Binding.. so it shares the same value????? idk when the form disappears 
i will update the user defaults or whatever if i compare changes with 
the old struct. how will i reference the old values? i will have to 
duplicate the struct i suppose, no need to read from disk twice..

fuck this project and fuck this entire work fuck fuck fuck fuck i am 
losing my mind with the sheer amount of hours i put into this project 
and every small incremental UX update is fucking complex to get just 
right and to go with a standardized theme to make the app feel cohesive 
is fucking rediculous