added alerts and verbose notifications
notification progress and all its changes is made very clear with alerts.

need to make custome alerts however because i want my UI images from previous implementation to be in the app.

still need a tutorial for each component so that the user can revisit and also digest on first launch!